chore: separate class and move to processors/ folder
