+++ packages/frontend/src/features/sqlRunner/components/visualizations/Table.tsx
@@ -1,106 +1,26 @@
-import { SqlRunnerResultsTransformer, type ResultRow } from '@lightdash/common';
+import { type ResultRow } from '@lightdash/common';
 import {
     flexRender,
     getCoreRowModel,
     useReactTable,
-    type ColumnDef,
-    type Table as TableType,
 } from '@tanstack/react-table';
-import { useVirtualizer, type Virtualizer } from '@tanstack/react-virtual';
+import { useVirtualizer } from '@tanstack/react-virtual';
 import { useRef, type FC } from 'react';
 import { SMALL_TEXT_LENGTH } from '../../../../components/common/LightTable';
 import BodyCell from '../../../../components/common/Table/ScrollableTable/BodyCell';
 import { VirtualizedArea } from '../../../../components/common/Table/ScrollableTable/TableBody';
 import {
-    ROW_HEIGHT_PX,
     Table as TableStyled,
     TableContainer,
     TableScrollableWrapper,
     TABLE_HEADER_BG,
     Tr,
 } from '../../../../components/common/Table/Table.styles';
-import { getRawValueCell } from '../../../../hooks/useColumns';
 import { type useSqlQueryRun } from '../../hooks/useSqlQueryRun';
-
-type TableChartSqlConfig =
-    | {
-          columns: Record<
-              string,
-              {
-                  visible: boolean;
-                  reference: string;
-                  label: string;
-                  frozen: boolean;
-                  order?: number;
-              }
-          >;
-      }
-    | undefined;
-
-class TableDataProcessor {
-    private transformer: SqlRunnerResultsTransformer;
-
-    private columns: ColumnDef<ResultRow, any>[];
-
-    private rows: ResultRow[];
-
-    constructor(
-        data: NonNullable<ReturnType<typeof useSqlQueryRun>['data']>,
-        private config: TableChartSqlConfig,
-    ) {
-        this.transformer = new SqlRunnerResultsTransformer({ data });
-        this.columns = this.createColumns();
-        this.rows = this.transformer.getRows();
-    }
-
-    private createColumns(): ColumnDef<ResultRow, any>[] {
-        const columns = this.transformer.getColumns();
-        return columns
-            .filter((column) =>
-                this.config ? this.config.columns[column]?.visible : true,
-            )
-            .map((column) => ({
-                id: column,
-                accessorKey: column,
-                header: this.config?.columns[column].label || column,
-                cell: getRawValueCell,
-            }));
-    }
-
-    public getColumns(): ColumnDef<ResultRow, any>[] {
-        return this.columns;
-    }
-
-    public getRows(): ResultRow[] {
-        return this.rows;
-    }
-
-    public getRowHeight(): number {
-        return ROW_HEIGHT_PX;
-    }
-
-    public getRowsCount(): number {
-        return this.rows.length;
-    }
-
-    public getColumnsCount(): number {
-        return this.columns.length;
-    }
-
-    public getTableData(
-        table: TableType<ResultRow>,
-        rowVirtualizer: Virtualizer<HTMLDivElement, Element>,
-    ) {
-        const { rows: rowModelRows } = table.getRowModel();
-        const virtualRows = rowVirtualizer.getVirtualItems();
-
-        return {
-            headerGroups: table.getHeaderGroups(),
-            virtualRows,
-            rowModelRows,
-        };
-    }
-}
+import {
+    TableDataProcessor,
+    type TableChartSqlConfig,
+} from '../../processors/TableDataProcessor';
 
 type Props = {
     data: NonNullable<ReturnType<typeof useSqlQueryRun>['data']>;
